Add label_name field to LC-MS schema
We have a "labeling" field that takes a boolean but if the user enters TRUE, we then need a place to describe the label. TMT is an LC-MS assay type that uses tandem mass tags to label individuals samples so they can be multiplexed. We're about to accept a large TMT collection off TMT datasets from Stanford.